refactor(money): normalize all money to integer CLP (Fase 1)
Dinero siempre en entero CLP (peso chileno, sin decimales) en toda la cadena.

Modelos (DecimalField -> PositiveIntegerField):
- Product.price, Product.compare_price (además arregla el tope de 9999.99 que
  impedía descuentos > $9.999)
- Order.amount, OrderItem.price
- Material.cost

Aritmética de dinero:
- carrito GetTotalView: suma como entero, sin float ni round(...,2)
- payment: unit_price / total_amount / transaction_amount como int (CLP)

Tests: total de carrito entero exacto, compare_price grande antes imposible,
Order.amount / OrderItem.price / Material.cost enteros.

+++ payment/views.py
@@ -63,8 +63,8 @@ class ProcessPaymentView(APIView):
                     productos['description'] = producto.description
                     productos['category_id'] = producto.category.id
                     productos['quantity'] = 1
-                    productos['unit_price'] = float(producto.price)
-                    total_amount += float(producto.price)
+                    productos['unit_price'] = int(producto.price)
+                    total_amount += int(producto.price)
                     items.append(productos)
             else:
                 return Response(
@@ -106,8 +106,8 @@ class ProcessPaymentView(APIView):
                     productos['description'] = producto.description
                     productos['category_id'] = producto.category.id
                     productos['quantity'] = 1
-                    productos['unit_price'] = float(producto.price)
-                    total_amount += float(producto.price)
+                    productos['unit_price'] = int(producto.price)
+                    total_amount += int(producto.price)
                     items.append(productos)
             else:
 
@@ -137,8 +137,8 @@ class ProcessPaymentView(APIView):
                         productos['description'] = str(product['description'])
                         productos['category_id'] = str(product['category'])
                         productos['quantity'] = 1
-                        productos['unit_price'] = float(product['price'])
-                        total_amount += float(product['price'])
+                        productos['unit_price'] = int(product['price'])
+                        total_amount += int(product['price'])
                         items.append(productos)
 
             
@@ -299,8 +299,8 @@ class PaymentsBricks(APIView):
                 productos['description'] = producto.description
                 productos['category_id'] = producto.category.id
                 productos['quantity'] = 1
-                productos['unit_price'] = float(producto.price)
-                total_amount += float(producto.price)
+                productos['unit_price'] = int(producto.price)
+                total_amount += int(producto.price)
                 items.append(productos)
         else:
 
@@ -329,8 +329,8 @@ class PaymentsBricks(APIView):
                     productos['description'] = str(product['description'])
                     productos['category_id'] = str(product['category'])
                     productos['quantity'] = 1
-                    productos['unit_price'] = float(product['price'])
-                    total_amount += float(product['price'])
+                    productos['unit_price'] = int(product['price'])
+                    total_amount += int(product['price'])
                     items.append(productos)
 
             
@@ -347,7 +347,7 @@ class PaymentsBricks(APIView):
             }
         
         payment_data = {
-            "transaction_amount": float(request.POST.get("transaction_amount")),
+            "transaction_amount": int(request.POST.get("transaction_amount")),
             "token": request.POST.get("token"),
             "description": request.POST.get("description"),
             "installments": int(request.POST.get("installments")),
